+++ ts/slicepkt.go
@@ -34,12 +34,12 @@ func (p SlicePkt) SetSync() {
 	p[0] = 0x47
 }
 
-func (p SlicePkt) Pid() uint16 {
-	return uint16(p[1]&0x1f)<<8 | uint16(p[2])
+func (p SlicePkt) Pid() int16 {
+	return int16(p[1]&0x1f)<<8 | int16(p[2])
 }
 
-func (p SlicePkt) SetPid(pid uint16) {
-	if pid > 8191 {
+func (p SlicePkt) SetPid(pid int16) {
+	if uint(pid) > 8191 {
 		panic("Bad PID")
 	}
 	p[1] = p[1]&0xe0 | byte(pid>>8)
